+++ capstonebundle/arm/arm32/common.h
@@ -14,7 +14,7 @@ void capstone_plugin_arm32_emulate(RDContext* ctx, const RDInstruction* instr,
                                    RDProcessor* p);
 
 bool capstone_plugin_arm32_render_operand(RDRenderer* r,
-                                          const RDInstruction* instr, usize idx,
+                                          const RDInstruction* instr, int idx,
                                           RDProcessor* p);
 
 bool capstone_arm32_decode_regmask(const cs_insn* cs_insn, RDInstruction* instr,
